+++ constants.ts
@@ -1,0 +1,10 @@
+export const safebooruRoot = "https://safebooru.donmai.us";
+// export const imagesPerGame = 9;
+// export const secondsBetweenImages = 3.0;
+// export const secondsBetweenHints = 30.0;
+export const imagesPerGame = 2;
+export const secondsBetweenImages = 2.0;
+export const secondsBetweenHints = 2.0;
+export const gameChannelNames = ["games"];
+export const topN = 10;
+export const tieSeconds = 0.5;
